Create a UI in script.js
This branch was created for a second version of the program where we
take it out of the console and give it a user interface to make it more
accessible and interactive. a lot of the old code was refactored, some
was made redundant and was completely changed.
Some decoration was performed on the UI but  was kept to a minimal
All the changes that were made to the code were perfromed solely using
js and DOM manipulation methods.